Fix cycle
