+++ packages/sharedClient/src/stores/FocusStore.ts
@@ -1,7 +1,8 @@
 import {create, type StateCreator} from 'zustand';
 
+import {DEFAULT_FOCUSED_NAV_ITEM} from '@shared/lib/navItems.shared';
+
 import type {FeedItemId} from '@shared/types/feedItems.types';
-import {DEFAULT_FOCUSED_NAV_ITEM} from '@shared/types/urls.types';
 import type {NavItemId} from '@shared/types/urls.types';
 import type {Func} from '@shared/types/utils.types';
